Remove the web-only /v2/validator API from the validator client
Deprecates `--write-wallet-password-on-web-onboarding` flag.

Also stop serving ALL endpoints start with `api.WebUrlPrefix` (== "/v2/validator") in the validator client.
Now validator client ONLY serves keymanager APIs.

VC RPC server now doesn't have to hold `healthClient` and `chainClient` anymore.

Note that `walletInitializedFeed` is also removed from VC RPC server
as it was only used in `initializeWallet` in `handler_wallet.go` which is now removed.

`AuthTokenHandler` gets much simpler, just check whether the path starts with keymanager api prefix
which is `"/eth/v1"`.
