+++ api/constants.go
@@ -1,9 +1,6 @@
 package api
 
 const (
-	WebUrlPrefix        = "/v2/validator/"
-	WebApiUrlPrefix     = "/api/v2/validator/"
 	KeymanagerApiPrefix = "/eth/v1"
-	SystemLogsPrefix    = "health/logs"
 	AuthTokenFileName   = "auth-token"
 )
