+++ config/features/deprecated_flags.go
@@ -65,6 +65,11 @@ var (
 		Usage:  deprecatedUsage,
 		Hidden: true,
 	}
+	deprecatedWriteWalletPasswordOnWebOnboarding = &cli.BoolFlag{
+		Name:   "write-wallet-password-on-web-onboarding",
+		Usage:  deprecatedUsage,
+		Hidden: true,
+	}
 )
 
 // Deprecated flags for both the beacon node and validator client.
@@ -78,6 +83,7 @@ var deprecatedFlags = []cli.Flag{
 	deprecatedInteropEth1DataVotes,
 	deprecatedInteropWriteSSZStateTransitions,
 	deprecatedWeb,
+	deprecatedWriteWalletPasswordOnWebOnboarding,
 }
 
 var upcomingDeprecation = []cli.Flag{
