--- io/logs/mock/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["mock_stream.go"],
-    importpath = "github.com/OffchainLabs/prysm/v7/io/logs/mock",
-    visibility = ["//visibility:public"],
-    deps = ["//async/event:go_default_library"],
-)
