+++ validator/rpc/BUILD.bazel
@@ -5,13 +5,7 @@ go_library(
     srcs = [
         "auth_token.go",
         "beacon.go",
-        "handler_wallet.go",
-        "handlers_accounts.go",
-        "handlers_auth.go",
-        "handlers_beacon.go",
-        "handlers_health.go",
         "handlers_keymanager.go",
-        "handlers_slashing.go",
         "handlers_validator_config.go",
         "intercepter.go",
         "log.go",
@@ -25,35 +19,21 @@ go_library(
     deps = [
         "//api:go_default_library",
         "//api/grpc:go_default_library",
-        "//api/pagination:go_default_library",
         "//api/rest:go_default_library",
-        "//api/server:go_default_library",
         "//api/server/httprest:go_default_library",
         "//api/server/middleware:go_default_library",
         "//api/server/structs:go_default_library",
-        "//async/event:go_default_library",
         "//beacon-chain/rpc/eth/shared:go_default_library",
-        "//cmd:go_default_library",
         "//cmd/validator/flags:go_default_library",
-        "//config/features:go_default_library",
         "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
         "//config/proposer:go_default_library",
         "//consensus-types/primitives:go_default_library",
         "//consensus-types/validator:go_default_library",
-        "//crypto/bls:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//io/file:go_default_library",
-        "//io/logs:go_default_library",
-        "//io/prompt:go_default_library",
         "//monitoring/tracing/trace:go_default_library",
         "//network/httputil:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "//runtime/version:go_default_library",
-        "//third_party/go-bip39:go_default_library",
-        "//third_party/go-bip39/wordlists:go_default_library",
-        "//validator/accounts:go_default_library",
-        "//validator/accounts/petnames:go_default_library",
         "//validator/accounts/wallet:go_default_library",
         "//validator/client:go_default_library",
         "//validator/client/iface:go_default_library",
@@ -61,7 +41,6 @@ go_library(
         "//validator/helpers:go_default_library",
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/derived:go_default_library",
-        "//validator/keymanager/local:go_default_library",
         "//validator/keymanager/remote-web3signer:go_default_library",
         "//validator/slashing-protection-history:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
@@ -74,7 +53,6 @@ go_library(
         "@com_github_grpc_ecosystem_go_grpc_prometheus//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_wealdtech_go_eth2_wallet_encryptor_keystorev4//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_protobuf//types/known/emptypb:go_default_library",
     ],
@@ -85,13 +63,7 @@ go_test(
     srcs = [
         "auth_token_test.go",
         "beacon_test.go",
-        "handler_wallet_test.go",
-        "handlers_accounts_test.go",
-        "handlers_auth_test.go",
-        "handlers_beacon_test.go",
-        "handlers_health_test.go",
         "handlers_keymanager_test.go",
-        "handlers_slashing_test.go",
         "handlers_validator_config_test.go",
         "intercepter_test.go",
         "server_test.go",
@@ -100,28 +72,22 @@ go_test(
     deps = [
         "//api:go_default_library",
         "//api/grpc:go_default_library",
-        "//async/event:go_default_library",
         "//cmd/validator/flags:go_default_library",
-        "//config/features:go_default_library",
         "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
         "//config/proposer:go_default_library",
         "//consensus-types/primitives:go_default_library",
         "//consensus-types/validator:go_default_library",
         "//crypto/bls:go_default_library",
-        "//crypto/rand:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//io/file:go_default_library",
-        "//io/logs/mock:go_default_library",
         "//network/httputil:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//testing/validator-mock:go_default_library",
-        "//third_party/go-bip39:go_default_library",
         "//validator/accounts:go_default_library",
         "//validator/accounts/iface:go_default_library",
-        "//validator/accounts/wallet:go_default_library",
         "//validator/client:go_default_library",
         "//validator/db/common:go_default_library",
         "//validator/db/filesystem:go_default_library",
@@ -135,12 +101,9 @@ go_test(
         "//validator/testing:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
-        "@com_github_golang_protobuf//ptypes/empty",
         "@com_github_google_uuid//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
         "@com_github_wealdtech_go_eth2_wallet_encryptor_keystorev4//:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//metadata:go_default_library",
         "@org_golang_google_protobuf//types/known/emptypb:go_default_library",
         "@org_golang_google_protobuf//types/known/timestamppb:go_default_library",
