+++ validator/rpc/intercepter.go
@@ -14,20 +14,10 @@ const bearerPrefix = "Bearer "
 // AuthTokenHandler is an HTTP handler to authorize a route.
 func (s *Server) AuthTokenHandler(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		path := r.URL.Path
-		needsAuth := strings.Contains(path, api.WebApiUrlPrefix) || strings.Contains(path, api.KeymanagerApiPrefix)
-		// Protect direct (non-/api) web endpoints too; otherwise callers can bypass auth by hitting /v2/validator/*.
-		if strings.HasPrefix(path, api.WebUrlPrefix) &&
-			!strings.HasPrefix(path, api.WebUrlPrefix+"initialize") &&
-			!strings.HasPrefix(path, api.WebUrlPrefix+"health/") {
-			needsAuth = true
-		}
-
-		if needsAuth && !strings.Contains(path, api.SystemLogsPrefix) {
-			// ignore some routes
+		if strings.HasPrefix(r.URL.Path, api.KeymanagerApiPrefix) {
 			reqToken := r.Header.Get("Authorization")
 			if reqToken == "" {
-				httputil.HandleError(w, "Unauthorized: no Authorization header passed. Please use an Authorization header with the jwt created in the prysm wallet", http.StatusUnauthorized)
+				httputil.HandleError(w, "Unauthorized: no Authorization header passed. Please use an Authorization header with the auth token created in the prysm wallet", http.StatusUnauthorized)
 				return
 			}
